+++ packages/database/drizzle/0007_giant_plazm.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "platform"."apps" ADD COLUMN "sidebar_component" varchar(100);